+++ randomgen/pcg32.pyx
@@ -305,6 +305,7 @@ cdef class PCG32:
           RNG.  For example, two 16-bit integer values can be simulated
           from a single draw of a 32-bit RNG.
         """
+        delta = wrap_int(delta, 64)
         pcg32_advance_state(&self.rng_state, <uint64_t>delta)
         return self
 
@@ -318,14 +319,22 @@ cdef class PCG32:
         ----------
         iter : integer, positive
             Number of times to jump the state of the rng.
+        
+        Notes
+        -----
+        The step size is phi when divided by the period 2**64
         """
-        self.advance(iter * 2**32)
+        step = int(0x9e3779b97f4a7c16)
+        self.advance(iter * step)
 
     def jump(self, iter=1):
         """
         jump(iter=1)
 
-        Jumps the state as-if 2**32 random numbers have been generated
+        Jump the state a fixed increment
+
+        Jumps the state as-if 11400714819323198486 random numbers have been
+        generated.
 
         Parameters
         ----------
@@ -336,6 +345,10 @@ cdef class PCG32:
         -------
         self : PCG32
             RNG jumped iter times
+
+        Notes
+        -----
+        The step size is phi when divided by the period 2**64
         """
         import warnings
         warnings.warn('jump (in-place) has been deprecated in favor of jumped'
@@ -351,7 +364,7 @@ cdef class PCG32:
         Returns a new bit generator with the state jumped
 
         The state of the returned big generator is jumped as-if
-        2**(32 * iter) random numbers have been generated.
+        11400714819323198486 random numbers have been generated.
 
         Parameters
         ----------
@@ -362,6 +375,10 @@ cdef class PCG32:
         -------
         bit_generator : PCG32
             New instance of generator jumped iter times
+
+        Notes
+        -----
+        The step size is phi when divided by the period 2**64            
         """
         cdef PCG32 bit_generator