+++ randomgen/philox.pyx
@@ -348,7 +348,7 @@ cdef class Philox:
         self.rng_state.uinteger = value['uinteger']
         self.rng_state.buffer_pos = value['buffer_pos']
 
-    cdef jump_inplace(self, np.npy_intp iter):
+    cdef jump_inplace(self, iter):
         """
         Jump state in-place
         
@@ -359,9 +359,9 @@ cdef class Philox:
         iter : integer, positive
             Number of times to jump the state of the rng.
         """
-        self.advance(iter * 2 ** 128)
+        self.advance(iter * int(2 ** 128))
 
-    def jump(self, np.npy_intp iter=1):
+    def jump(self, iter=1):
         """
         jump(iter=1)
 
@@ -451,6 +451,8 @@ cdef class Philox:
         Advancing the RNG state resets any pre-computed random numbers.
         This is required to ensure exact reproducibility.
         """
+        delta = wrap_int(delta, 256)
+
         cdef np.ndarray delta_a
         delta_a = int_to_array(delta, 'step', 256, 64)
         philox_advance(<uint64_t *>np.PyArray_DATA(delta_a), &self.rng_state)