+++ randomgen/threefry.pyx
@@ -352,9 +352,9 @@ cdef class ThreeFry:
         iter : integer, positive
             Number of times to jump the state of the rng.
         """
-        self.advance(iter * 2 ** 128)
+        self.advance(iter * int(2**128))
 
-    def jump(self, np.npy_intp iter=1):
+    def jump(self, iter=1):
         """
         jump(iter=1)
 
@@ -381,7 +381,7 @@ cdef class ThreeFry:
         self.jump_inplace(iter)
         return self
 
-    def jumped(self, np.npy_intp iter=1):
+    def jumped(self, iter=1):
         """
         jumped(iter=1)
 
@@ -444,6 +444,8 @@ cdef class ThreeFry:
         Advancing the RNG state resets any pre-computed random numbers.
         This is required to ensure exact reproducibility.
         """
+        delta = wrap_int(delta, 256)
+
         cdef np.ndarray delta_a
         delta_a = int_to_array(delta, 'step', 256, 64)
         threefry_advance(<uint64_t *>np.PyArray_DATA(delta_a), &self.rng_state)